+++ backend/src/controllers/notes.controller.ts
@@ -1,93 +1,85 @@
 import { Response } from 'express';
 import { AuthenticatedRequest } from '../models';
-import {cardsService, notesService} from '../services';
-import { PrismaClient } from '@prisma/client';
+import { notesService} from '../services';
+import {BaseController} from "./base.controller";
 
-export class NotesController {
-    #prisma!: PrismaClient;
-
-    constructor() {
-        this.#prisma = new PrismaClient();
-    }
-
+export class NotesController extends BaseController {
     public async getNotes(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
-            const userId = req.user?.id;
+            const userId = this.extractAuthenticatedUserId(req, res);
 
             if (!userId) {
-                return Promise.reject(new Error('User not authenticated'));
+                return;
             }
 
             const notes = await notesService.getNotesByUserId(userId);
 
-            if (!notes) {
-                res.status(404).json({ error: 'Notes not found' });
-            } else {
-                res.status(200).json(notes);
-            }
+            res.status(200).json(notes);
         }
         catch (error) {
-            res.status(500).json({ error: 'Failed to fetch notes' });
+            res.status(500).json({ error: 'Internal server error' });
         }
 
     }
 
     public async getNoteById(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
-            const userId = req.user?.id;
+            const userId = this.extractAuthenticatedUserId(req, res);
 
             if (!userId) {
-                return Promise.reject(new Error('User not authenticated'));
+                return;
             }
 
-            const noteId = parseInt(req.params.id, 10);
+            const noteId = this.getValidatedParameterValue(req, res);
+
+            if (!noteId) {
+                return;
+            }
+
             const note = await notesService.getNoteById(userId, noteId);
 
-            if (!note) {
-                res.status(404).json({ error: 'Note not found' });
-            } else {
-                res.status(200).json(note);
-            }
+            res.status(200).json(note);
         }
         catch (error) {
-            res.status(500).json({ error: 'Failed to fetch note' });
+            res.status(500).json({ error: 'Internal server error' });
         }
     }
 
     public async createNote(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
-            const userId = req.user?.id;
+            const userId = this.extractAuthenticatedUserId(req, res);
 
             if (!userId) {
-                return Promise.reject(new Error('User not authenticated'));
+                return;
             }
 
             const note = req.body;
-            const createdNote = notesService.createNote(userId, note);
+            const createdNote = await notesService.createNote(userId, note);
 
             res.status(201).json(createdNote);
         } catch (error) {
-            res.status(500).json({ error: 'Failed to create note' });
+            res.status(500).json({ error: 'Internal server error' });
         }
     }
 
     public async updateNote(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
-            const userId = req.user?.id;
+            const userId = this.extractAuthenticatedUserId(req, res);
 
             if (!userId) {
-                return Promise.reject(new Error('User not authenticated'));
+                return;
             }
 
-            const noteId = parseInt(req.params.id, 10);
+            const noteId = this.getValidatedParameterValue(req, res);
+
+            if (!noteId) {
+                return;
+            }
+
             const updatedNote = req.body;
             const note = await notesService.updateNote(userId, noteId, updatedNote);
 
-            if (!note) {
-                res.status(404).json({ error: 'Note not found' });
-            }
-
-            res.status(200).json({ message: 'Note updated successfully' });
+            res.status(200).json({ message: 'Note updated successfully', note });
         } catch (error) {
             res.status(500).json({ error: 'Internal server error' });
         }
@@ -95,23 +87,24 @@ export class NotesController {
 
     public async deleteNote(req: AuthenticatedRequest, res: Response): Promise<void> {
         try {
-            const userId = req.user?.id;
+            const userId = this.extractAuthenticatedUserId(req, res);
 
             if (!userId) {
-                return Promise.reject(new Error('User not authenticated'));
+                return;
             }
 
-            const noteId = parseInt(req.params.id, 10);
+            const noteId = this.getValidatedParameterValue(req, res);
+
+            if (!noteId) {
+                return;
+            }
+
             const deletedNote = await notesService.deleteNote(userId, noteId);
 
-            if (!deletedNote) {
-                res.status(404).json({ error: 'Note not found' });
-            } else {
-                res.status(200).json({ message: 'Note deleted successfully' });
-            }
+            res.status(200).json({ message: 'Note deleted successfully', note: deletedNote });
         }
         catch (error) {
-            res.status(500).json({ error: 'Failed to delete note' });
+            res.status(500).json({ error: 'Internal server error' });
         }
     }
 }
